+++ tests/ApcuStorageTest.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace Yiisoft\Yii\RateLimiter\Tests;
 
 use InvalidArgumentException;
-use Yiisoft\Yii\RateLimiter\Exception\CannotUseException;
 use Yiisoft\Yii\RateLimiter\Storage\ApcuStorage;
 use Yiisoft\Yii\RateLimiter\Storage\StorageInterface;
 
@@ -31,13 +30,63 @@ final class ApcuStorageTest extends StorageTest
         }
     }
 
+    public function testSaveIfNotExistsWithExistsKey(): void
+    {
+        $storage = $this->getStorage();
+
+        $value = (new FrozenTimeTimer())->nowInMilliseconds();
+        $storage->saveIfNotExists('exists_key', $value, self::DEFAULT_TTL);
+
+        $result = $storage->saveIfNotExists('exists_key', $value, self::DEFAULT_TTL);
+
+        $this->assertFalse($result);
+    }
+
+    public function testSaveCompareAndSwapWithNewKey(): void
+    {
+        $storage = $this->getStorage();
+
+        $newValue = (new FrozenTimeTimer())->nowInMilliseconds();
+        $oldValue = (int) $storage->get('new_key');
+
+        $result = $storage->saveCompareAndSwap(
+            'new_key', 
+            $oldValue, 
+            $newValue, 
+            self::DEFAULT_TTL
+        );
+
+        $this->assertFalse($result);
+    }
+
+    public function testSaveCompareAndSwapWithExistsKeyButOldValueDiffrent(): void
+    {
+        $storage = $this->getStorage();
+
+        $oldValue = (new FrozenTimeTimer())->nowInMilliseconds();
+        $storage->saveIfNotExists('exists_key', $oldValue, self::DEFAULT_TTL);
+
+        $oldValue = $oldValue + 200;
+
+        $newValue = $oldValue + 100;
+
+        $result = $storage->saveCompareAndSwap(
+            'exists_key', 
+            $oldValue, 
+            $newValue, 
+            self::DEFAULT_TTL
+        );
+
+        $this->assertFalse($result);
+    }
+
     public function testInvalidArgumentExceptionWithGet(): void
     {
-        $storage = new ApcuStorage();
-
         apcu_add('key', 'string_value', parent::DEFAULT_TTL);
 
         $this->expectException(InvalidArgumentException::class);
+
+        $storage = $this->getStorage();
         $storage->get('key');
     }
 }
